finalize spec
